# grpc/api/openapiv2/route_guide/route_guide.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "route_guide/route_guide.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "RouteGuide"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/routeguide.RouteGuide/GetFeature": {
      "post": {
        "summary": "A simple RPC.",
        "description": "Obtains the feature at a given position.\n\nA feature with an empty name is returned if there's no feature at the given\nposition.",
        "operationId": "RouteGuide_GetFeature",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/routeguide.Feature"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/google.rpc.Status"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "Points are represented as latitude-longitude pairs in the E7 representation\n(degrees multiplied by 10**7 and rounded to the nearest integer).\nLatitudes should be in the range +/- 90 degrees and longitude should be in\nthe range +/- 180 degrees (inclusive).",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/routeguide.Point"
            }
          }
        ],
        "tags": [
          "RouteGuide"
        ]
      }
    },
    "/routeguide.RouteGuide/ListFeatures": {
      "post": {
        "summary": "A server-to-client streaming RPC.",
        "description": "Obtains the Features available within the given Rectangle.  Results are\nstreamed rather than returned at once (e.g. in a response message with a\nrepeated field), as the rectangle may cover a large area and contain a\nhuge number of features.",
        "operationId": "RouteGuide_ListFeatures",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/routeguide.Feature"
                },
                "error": {
                  "$ref": "#/definitions/google.rpc.Status"
                }
              },
              "title": "Stream result of routeguide.Feature"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/google.rpc.Status"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "A latitude-longitude rectangle, represented as two diagonally opposite\npoints \"lo\" and \"hi\".",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/routeguide.Rectangle"
            }
          }
        ],
        "tags": [
          "RouteGuide"
        ]
      }
    },
    "/routeguide.RouteGuide/RecordRoute": {
      "post": {
        "summary": "A client-to-server streaming RPC.",
        "description": "Accepts a stream of Points on a route being traversed, returning a\nRouteSummary when traversal is completed.",
        "operationId": "RouteGuide_RecordRoute",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/routeguide.RouteSummary"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/google.rpc.Status"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "Points are represented as latitude-longitude pairs in the E7 representation\n(degrees multiplied by 10**7 and rounded to the nearest integer).\nLatitudes should be in the range +/- 90 degrees and longitude should be in\nthe range +/- 180 degrees (inclusive). (streaming inputs)",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/routeguide.Point"
            }
          }
        ],
        "tags": [
          "RouteGuide"
        ]
      }
    },
    "/routeguide.RouteGuide/RouteChat": {
      "post": {
        "summary": "A Bidirectional streaming RPC.",
        "description": "Accepts a stream of RouteNotes sent while a route is being traversed,\nwhile receiving other RouteNotes (e.g. from other users).",
        "operationId": "RouteGuide_RouteChat",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/routeguide.RouteNote"
                },
                "error": {
                  "$ref": "#/definitions/google.rpc.Status"
                }
              },
              "title": "Stream result of routeguide.RouteNote"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/google.rpc.Status"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "A RouteNote is a message sent while at a given point. (streaming inputs)",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/routeguide.RouteNote"
            }
          }
        ],
        "tags": [
          "RouteGuide"
        ]
      }
    }
  },
  "definitions": {
    "google.protobuf.Any": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "google.rpc.Status": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/google.protobuf.Any"
          }
        }
      }
    },
    "routeguide.Feature": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "description": "The name of the feature."
        },
        "location": {
          "$ref": "#/definitions/routeguide.Point",
          "description": "The point where the feature is detected."
        }
      },
      "description": "A feature names something at a given point.\n\nIf a feature could not be named, the name is empty."
    },
    "routeguide.Point": {
      "type": "object",
      "properties": {
        "latitude": {
          "type": "integer",
          "format": "int32"
        },
        "longitude": {
          "type": "integer",
          "format": "int32"
        }
      },
      "description": "Points are represented as latitude-longitude pairs in the E7 representation\n(degrees multiplied by 10**7 and rounded to the nearest integer).\nLatitudes should be in the range +/- 90 degrees and longitude should be in\nthe range +/- 180 degrees (inclusive)."
    },
    "routeguide.Rectangle": {
      "type": "object",
      "properties": {
        "lo": {
          "$ref": "#/definitions/routeguide.Point",
          "description": "One corner of the rectangle."
        },
        "hi": {
          "$ref": "#/definitions/routeguide.Point",
          "description": "The other corner of the rectangle."
        }
      },
      "description": "A latitude-longitude rectangle, represented as two diagonally opposite\npoints \"lo\" and \"hi\"."
    },
    "routeguide.RouteNote": {
      "type": "object",
      "properties": {
        "location": {
          "$ref": "#/definitions/routeguide.Point",
          "description": "The location from which the message is sent."
        },
        "message": {
          "type": "string",
          "description": "The message to be sent."
        }
      },
      "description": "A RouteNote is a message sent while at a given point."
    },
    "routeguide.RouteSummary": {
      "type": "object",
      "properties": {
        "pointCount": {
          "type": "integer",
          "format": "int32",
          "description": "The number of points received."
        },
        "featureCount": {
          "type": "integer",
          "format": "int32",
          "description": "The number of known features passed while traversing the route."
        },
        "distance": {
          "type": "integer",
          "format": "int32",
          "description": "The distance covered in metres."
        },
        "elapsedTime": {
          "type": "integer",
          "format": "int32",
          "description": "The duration of the traversal in seconds."
        }
      },
      "description": "A RouteSummary is received in response to a RecordRoute rpc.\n\nIt contains the number of individual points received, the number of\ndetected features, and the total distance covered as the cumulative sum of\nthe distance between each point."
    }
  }
}